- Update password handling validation and frontend views based on existence of user's password
