+++ app/Http/Controllers/ProfileController.php
@@ -23,7 +23,7 @@ class ProfileController extends Controller
         return Inertia::render('Profile/Edit', [
             'mustVerifyEmail' => $request->user() instanceof MustVerifyEmail,
             'status' => session('status'),
-            'mustVerifyPasswordToDeleteAccount' => $request->user()->password !== null,
+            'hasPassword' => $request->user()->password !== null,
         ]);
     }
