+++ app/Http/Requests/PasswordUpdateRequest.php
@@ -1,0 +1,42 @@
+<?php
+
+namespace App\Http\Requests;
+
+use Illuminate\Foundation\Http\FormRequest;
+use Illuminate\Validation\Rules\Password;
+
+class PasswordUpdateRequest extends FormRequest
+{
+    public function authorize(): bool
+    {
+        return true;
+    }
+
+    public function rules(): array
+    {
+        $user = $this->user();
+
+        if ($user->password) {
+            // User is updating an existing password
+            return [
+                'current_password' => ['required', 'current_password'],
+                'password' => ['required', Password::defaults(), 'confirmed'],
+            ];
+        } else {
+            // User is setting a password for the first time
+            return [
+                'password' => ['required', Password::defaults(), 'confirmed'],
+            ];
+        }
+    }
+
+    public function messages(): array
+    {
+        return [
+            'current_password.required' => 'Please enter your current password.',
+            'current_password.current_password' => 'The current password is incorrect.',
+            'password.required' => 'Please enter a new password.',
+            'password.confirmed' => 'The password confirmation does not match.',
+        ];
+    }
+}
